Fix "circular dependency" by moving `printTypeAnnotation` to `misc.js`
