+++ src/language-js/print/array.js
@@ -8,8 +8,7 @@ const { getLast, isNextLineEmpty } = require("../../common/util");
 const { hasDanglingComments, shouldPrintComma } = require("../utils");
 const { locEnd } = require("../loc");
 
-const { printOptionalToken } = require("./misc");
-const { printTypeAnnotation } = require("./type-annotation");
+const { printOptionalToken, printTypeAnnotation } = require("./misc");
 
 /** @typedef {import("../../document").Doc} Doc */
