+++ src/language-js/print/misc.js
@@ -3,6 +3,7 @@
 const {
   builders: { concat, indent, join, line },
 } = require("../../document");
+const { isFlowAnnotationComment } = require("../utils");
 
 function printOptionalToken(path) {
   const node = path.getValue();
@@ -34,6 +35,32 @@ function printFunctionTypeParameters(path, options, print) {
   return "";
 }
 
+function printTypeAnnotation(path, options, print) {
+  const node = path.getValue();
+  if (!node.typeAnnotation) {
+    return "";
+  }
+
+  const parentNode = path.getParentNode();
+  const isDefinite =
+    node.definite ||
+    (parentNode &&
+      parentNode.type === "VariableDeclarator" &&
+      parentNode.definite);
+
+  const isFunctionDeclarationIdentifier =
+    parentNode.type === "DeclareFunction" && parentNode.id === node;
+
+  if (isFlowAnnotationComment(options.originalText, node.typeAnnotation)) {
+    return concat([" /*: ", path.call(print, "typeAnnotation"), " */"]);
+  }
+
+  return concat([
+    isFunctionDeclarationIdentifier ? "" : isDefinite ? "!: " : ": ",
+    path.call(print, "typeAnnotation"),
+  ]);
+}
+
 function printBindExpressionCallee(path, options, print) {
   return concat(["::", path.call(print, "callee")]);
 }
@@ -63,5 +90,6 @@ module.exports = {
   printFunctionTypeParameters,
   printBindExpressionCallee,
   printTypeScriptModifiers,
+  printTypeAnnotation,
   adjustClause,
 };
